scrap/requirements.txt
unidecode 
requests 
pandas
datetime 
google-cloud-storage